+++ Gruntfile.js
@@ -9,7 +9,10 @@ module.exports = function (grunt) {
             { src: "lib/services/Contacts.js", dest: "build/contacts.md" },
             { src: "lib/services/IM.js", dest: "build/im.md" },
             { src: "lib/services/Presence.js", dest: "build/presence.md" },
-            { src: "index.js", dest: "build/sdk.md" }
+            { src: "index.js", dest: "build/sdk.md" },
+            { src: "lib/common/models/Bubble.js", dest: "build/bubble.md" },
+            { src: "lib/common/models/Contact.js", dest: "build/contact.md" },
+            { src: "lib/common/models/Message.js", dest: "build/message.md" },
         ]
       }
     },